+++ src/Translator/FraudStatus.php
@@ -6,33 +6,41 @@ class FraudStatus
 {
     const ACCEPT = 'accept';
 
-    const REJECT = 'reject';
+    const DENY = 'deny';
 
-    public function translate(string $status)
+    const CHALLENGE = 'challenge';
+
+    public function translate(string $status): array
     {
         return match ($status) {
             self::ACCEPT => $this->accept(),
-            self::REJECT => $this->reject(),
+            self::DENY => $this->deny(),
+            self::CHALLENGE => $this->challenge(),
             default => $this->respond('Invalid fraud status', 400),
         };
     }
 
-    private function respond(string $message, int $code = 200)
+    private function respond(string $message, int $code = 200): array
     {
-        return response()->json([
+        return [
             'code' => $code,
             'status' => $code === 200 ? 'success' : 'error',
             'message' => $message,
-        ], $code);
+        ];
     }
 
-    private function accept()
+    private function accept(): array
     {
-        return $this->respond('Transaction is safe to proceed. its not considered as fraud.');
+        return $this->respond('Transaction is safe to proceed. Its not considered as fraud.');
     }
 
-    private function reject()
+    private function deny(): array
     {
-        return $this->respond('Transaction is considered as fraud. it is rejected by Midtrans.');
+        return $this->respond('Transaction is considered as fraud. It is denied by Midtrans.');
+    }
+
+    private function challenge(): array
+    {
+        return $this->respond('Transaction is flagged as challenge by Midtrans Fraud Detection System. Approve or deny the transaction.');
     }
 }