+++ tests/Pest.php
@@ -1,0 +1,5 @@
+<?php
+
+use BlissJaspis\Midtrans\Tests\TestCase;
+
+pest()->extend(TestCase::class)->in('Unit');